rpi-eink-test/requirements.txt
# e-Paper 스크립트 공통
Pillow>=9.0.0
pyserial>=3.5
